step4 refactor : parser -> converter로 이름 변경
