+++ src/main/kotlin/racinggame/view/converter/DefaultConverter.kt
@@ -1,0 +1,5 @@
+package racinggame.view.converter
+
+class DefaultConverter<T> : InputConverter<T> {
+    override fun convert(input: String): T = input as T
+}
