+++ src/main/kotlin/racinggame/view/converter/InputConverter.kt
@@ -1,0 +1,5 @@
+package racinggame.view.converter
+
+interface InputConverter<R> {
+    fun convert(input: String): R
+}
