+++ src/main/kotlin/racinggame/view/converter/NameConverter.kt
@@ -1,0 +1,11 @@
+package racinggame.view.converter
+
+class NameConverter : InputConverter<List<String>> {
+    override fun convert(input: String): List<String> {
+        return input.split(DELIMITER)
+    }
+
+    companion object {
+        private const val DELIMITER = ","
+    }
+}
