--- src/main/kotlin/racinggame/view/parser/DefaultParser.kt
@@ -1,5 +1,0 @@
-package racinggame.view.parser
-
-class DefaultParser<T> : InputParser<T> {
-    override fun parse(input: String): T = input as T
-}
