--- src/main/kotlin/racinggame/view/parser/InputParser.kt
@@ -1,5 +1,0 @@
-package racinggame.view.parser
-
-interface InputParser<R> {
-    fun parse(input: String): R
-}
